+++ Source/Plugins/UDBScript/API/ImageInfo.cs
@@ -28,7 +28,7 @@ using CodeImp.DoomBuilder.UDBScript.Wrapper;
 
 #endregion
 
-namespace CodeImp.DoomBuilder.UDBScript.API
+namespace CodeImp.DoomBuilder.UDBScript.Wrapper
 {
 	struct ImageInfo
 	{